requirements.txt
pyyaml>=5.1.0
typing>=3.7.4.3